Sketch of lazy tx state
This is a rough sketch of how a lazy tx state could work. Lazy in this
case means that updates to the txstate are only applied when needed,
ei. on read operations. Actual speedup happens if updates are not force
until commit/rollback, in which case index entry update computations can
be discarded. These can be pretty expensive especially for composite indexes,
so this could save some time.