Initial commit of vectorized, primitive enabled runtime prototype
This prototype processes batches of rows in a data structure refered to
as a Morsel. This structure actually stores rows by column, and allows
to specialize columns where the value type is known with enough certainty.

The Executable classes are executable operators which executes on an input
Morsel, and then pushes the Morsel to the next operator.