Fix and optimize copying of list-comprehensions
We want to copy the execution context to isolate execution inside
the list comprehensions from the rest of the query. However, we do
not need to copy on every element, but only once before the execution
starts. The same context can then be reused for all the elements, but
with an updated element variable.