Remove NodeProxyWrappingValue
Nodes are now passed using NodeReference by default, and materialized in ProduceResults
before being returned. In this commit, I also renamed

  NodeValue -> NodeFullValue
  VirtualNodeValue -> NodeValue

This means that the most general node value is now called NodeValue, and is used to
match for nodes all over Cypher. The hack for the db.schema procedure to be able to
return it's results as a graph is moves to the special DBSchemaProcedureNode class
which implements both NodeValue and org.neo4j.graphdb.Node.