Experimental Bolt adaptation to new runtime API
In a quick spike, the Bolt Server has be changed to use
the new execute method, which takes a result buffer, and
returns a QueryExecution. There is much more clean-up work
to do in Bolt, and more optimizations and implementations
to build in Cypher.

Still, some simple cypher queries in AuthenticationIT
work, so the happy path seems solved.