Experimental Core API adaptation to new runtime API
In another awesome hack, the Core API in changed to use the new
runtime API and a result buffer. Note that the Core API has to
back-pressure on every row to convert the push-based runtime API
to serve rows in a controlled way to the Core API Result iterator.

We still have a lot of connecting tissue to add around execution
plans, statistics and more, but the primary data concept seems
to work.